:wrench: PostCSS config updates due to API changes in latest postcss-loader version
