+++ webpack.config.js
@@ -1,9 +1,5 @@
 require('dotenv').config();
-const autoprefixer = require('autoprefixer');
-const cssnano = require('cssnano');
 const path = require('path');
-const purgecss = require('@fullhuman/postcss-purgecss');
-const tailwindcss = require('tailwindcss');
 const webpack = require('webpack');
 const CompressionPlugin = require('compression-webpack-plugin');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
@@ -14,69 +10,7 @@ const WorkboxPlugin = require('workbox-webpack-plugin');
 
 const mode = process.env.NODE_ENV;
 
-const cssWhitelistClassArray = [
-  /tippy/,
-  /leaflet/,
-  /leaflet-container/,
-  /leaflet-controls-container/,
-  /pill/,
-  /orange/,
-  /red/,
-  /yellow/,
-  /green/,
-  /purple/,
-  /bg-blue-400/,
-  /bg-blue-500/,
-  /bg-purple-400/,
-  /bg-purple-500/,
-  /opacity-75/,
-  /bg-gray-200/,
-  /bg-gray-300/,
-  /bg-gray-400/,
-  /bg-gray-500/,
-  /bg-gray-600/,
-  /bg-white/,
-];
-
-const postCssPluginsArray = [
-  autoprefixer(),
-  tailwindcss(),
-  cssnano({
-    preset: 'default',
-  }),
-];
-if (mode === 'production') {
-  postCssPluginsArray.push(
-    purgecss({
-      content: [
-        './public/index.html',
-        './src/components/**/*.js',
-        './src/components/**/*.jsx',
-      ],
-      defaultExtractor: (content) => {
-        // Capture as liberally as possible, including things like `h-(screen-1.5)`
-        const broadMatches = content.match(/[^<>"'`\s]*[^<>"'`\s:]/g) || [];
-        // Capture classes within other delimiters like .block(class="w-1/2") in Pug
-        const innerMatches = content.match(/[^<>"'`\s.()]*[^<>"'`\s.():]/g) || [];
-        return broadMatches.concat(innerMatches);
-      },
-      fontFace: false,
-      whitelistPatterns: cssWhitelistClassArray,
-      whitelistPatternsChildren: cssWhitelistClassArray,
-    }),
-  );
-}
-
 const webpackRules = [
-  // {
-  //   test: /\.(ttf|eot|woff|woff2)$/,
-  //   use: {
-  //     loader: 'file-loader',
-  //     options: {
-  //       name: 'fonts/[name].[ext]',
-  //     },
-  //   },
-  // },
   {
     test: /\.(sa|sc|c)ss$/,
     exclude: [/old/],
@@ -91,12 +25,6 @@ const webpackRules = [
       },
       {
         loader: 'postcss-loader',
-        options: {
-          sourceMap: true,
-          plugins() {
-            return postCssPluginsArray;
-          },
-        },
       },
       {
         loader: 'sass-loader',
@@ -133,16 +61,6 @@ const webpackPlugins = [
       },
     ],
   }),
-  // new CopyWebpackPlugin({
-  //   patterns: [
-  //     {
-  //       from: './public/fonts/*.woff2',
-  //       to: './css/fonts',
-  //       flatten: true,
-  //       force: true,
-  //     },
-  //   ],
-  // }),
   new CopyWebpackPlugin({
     patterns: [
       {
